Sanitizing input to patch path traversal flows

+++ src/main/java/org/owasp/webgoat/webwolf/FileServer.java
@@ -89,7 +89,14 @@ public class FileServer {
     // https://stackoverflow.com/questions/60336929/java-nio-file-nosuchfileexception-when-file-transferto-is-called
     try (InputStream is = multipartFile.getInputStream()) {
       var originalFilename = multipartFile.getOriginalFilename(); //Added IT24100500 for test
-      var destinationFile = new File(destinationDir, originalFilename);
+      var destinationFile = destinationDir.toPath().resolve(originalFilename).normalize();
+      if (!destinationFile.getParent().equals(destinationDir.toPath().toAbsolutePath())
+          && !destinationFile.getParent().equals(destinationDir.toPath())) {
+        log.warn("Rejected upload with path component from {}", username);
+        return new ModelAndView(
+            new RedirectView("files", true),
+            new ModelMap().addAttribute("uploadSuccess", NOTHING_TO_UPLOAD));
+      }
       Files.deleteIfExists(destinationFile);
       Files.copy(is, destinationFile);
     }
